Move most egui components to files
